local_jobs/requirements.txt
boto3>=1.26.0
pyspark>=3.1.0,<4.0.0
delta-spark>=3.2.0
